Move plugins listener from global subscribers into plugins
